Cleaned up custom inspector
