+++ Assets/Scripts/Editor/CustomInspector.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 023b2dd05b7841d49b67970f6046ebef
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
